requirements.txt
numpy
opencv-python
praw
python-dotenv
tinydb
python-twitter
